requirements.txt
django-bootstrap4==0.0.6
django-filter==1.1.0
django-tables2==1.21.2
django==2.0.4
pytz==2018.3
